Move interop files into more appropriate files
