+++ src/ClassiCube.vcxproj.filters
@@ -764,10 +764,27 @@
     <ClCompile Include="Window_Terminal.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Platform_MacClassic.c">
+      <Filter>Source Files\Platform</Filter>
+    </ClCompile>
+    <ClCompile Include="Window_MacClassic.c">
+      <Filter>Source Files\Window</Filter>
+    </ClCompile>
+    <ClCompile Include="Window_BeOS.cpp">
+      <Filter>Source Files\Window</Filter>
+    </ClCompile>
+    <ClCompile Include="Platform_BeOS.cpp">
+      <Filter>Source Files\Platform</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\misc\windows\CCicon.rc">
       <Filter>Resource Files</Filter>
     </ResourceCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Window_cocoa.m">
+      <Filter>Source Files\Window</Filter>
+    </None>
   </ItemGroup>
 </Project>
